LayoutTests/platform/ios/fast/dom/focus-dialog-blur-input-type-change-crash-expected.txt

CONSOLE MESSAGE: RangeError: Maximum call stack size exceeded.
PASS
